BPF: Simplify by splitting OS specifics into own files
